Create structure for Devstack plugin
This adds the basic structure of the Devstack plugin. Required to
enable a gate which actually runs this plugin. Patch to follow that
sets up the and configures the mixmatch project in Devstack.

It also enable bashate in the pep8 job for the plugin.sh file.
Bashate is comparable to pep8, but for bash scripts.

Change-Id: I142def16a3179ec5cc390f419e2442dd3b93683c