uwsgi improvements
This patch adds a .ini file with the configuration for running
mixmatch through uwsgi, replacing the previous `run_proxy.sh`
script which passed the configuration as arguments.

This patch also removed the wsgi configuration, making it
unsupported. The wsgi documentation has also been removed. A
follow on patch will also remove any remnants of wsgi logic.

The default port for running the proxy has been changed to
9913.

This patch also updates the `setup.cfg` file, adding the
parameters for wsgi_scripts so that pbr install the wsgi exec
on package installation.

Change-Id: I2069d0f28e8e14da59109158cd2912ddb284a4e3