Multiple Log files
Multiple loggers are created in
mixmatch/config/__init__.py. In this module,
there is a function called LOG that will determine
which logger to assign. To implemenet this, I
had to change the parameters of functions
to take services in order to use the LOG
function. I still need to add an output
file for every logger and debug from
unit tests. I wanted to
show what I currently have.

Change-Id: I0083eed093436f16aa28400a6f1652bbff6ff579