[WIP]Extend-a-network and mangle-a-subnet
Implemented as an extension, so a lot of stuff feels hacky. It's also
very light on exception handling. There are a lot of ways to break it,
but a lot of the issues will probably go away when the logic is ported
into a new service.

The subnet mangling feature either needs enhancement or removal.

Has not been tested with actual K2K auth.

Change-Id: I9088e3509f71fb363ddc7f504cbb96f94932cc1e